+++ src/main/java/chess/repository/dao/PieceDao.java
@@ -23,12 +23,12 @@ public class PieceDao implements PieceRepository {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public void savePieces(ChessGameManager chessGameManager, int gameId) {
+    public void savePieces(ChessGameManager chessGameManager, long gameId) {
         Map<String, PieceDto> board = ChessBoardDto.from(chessGameManager.getBoard()).board();
         String query = "INSERT INTO piece(game_id, name, color, position) VALUES(?, ?, ?, ?)";
         Set<String> positions = board.keySet();
         this.jdbcTemplate.batchUpdate(query, positions, positions.size(), (ps, argument) -> {
-            ps.setInt(1, gameId);
+            ps.setLong(1, gameId);
             PieceDto piece = board.get(argument);
             ps.setString(2, piece.getName());
             ps.setString(3, piece.getColor());
@@ -36,7 +36,7 @@ public class PieceDao implements PieceRepository {
         });
     }
 
-    public ChessBoard findChessBoardByGameId(int gameId) {
+    public ChessBoard findChessBoardByGameId(long gameId) {
         String queryPiece = "SELECT * FROM piece WHERE game_id = ?";
         ChessBoard chessBoard = this.jdbcTemplate.queryForObject(queryPiece, chessBoardRowMapper, gameId);
         return chessBoard;
@@ -56,7 +56,7 @@ public class PieceDao implements PieceRepository {
         return ChessBoard.from(board);
     };
 
-    public Piece findPieceByPosition(Position position, int gameId) {
+    public Piece findPieceByPosition(Position position, long gameId) {
         String queryPieceByPosition = "SELECT * FROM piece WHERE game_id = ? AND position = ?";
         Piece piece = this.jdbcTemplate.queryForObject(queryPieceByPosition, pieceRowMapper, gameId, position.getNotation());
         return piece;
@@ -68,18 +68,18 @@ public class PieceDao implements PieceRepository {
                     Color.of(resultSet.getString("color"))
             );
 
-    public void savePiece(Piece piece, Position position, int gameId) {
+    public void savePiece(Piece piece, Position position, long gameId) {
         String query = "INSERT INTO piece(game_id, name, color, position) VALUES(?, ?, ?, ?)";
         this.jdbcTemplate.update(query, gameId, piece.getName(), piece.getColor().name(), position.getNotation());
     }
 
-    public void deletePieceByPosition(Position position, int gameId) {
+    public void deletePieceByPosition(Position position, long gameId) {
         String queryPieceByPosition = "DELETE FROM piece WHERE game_id = ? AND position = ?";
         this.jdbcTemplate.update(queryPieceByPosition, gameId, position.getNotation());
     }
 
     @Override
-    public void updatePiecePosition(Position fromPosition, Position toPosition, int gameId) {
+    public void updatePiecePosition(Position fromPosition, Position toPosition, long gameId) {
         String query = "UPDATE piece SET position = ? WHERE game_id = ? AND position = ?";
         this.jdbcTemplate.update(query, toPosition.getNotation(), gameId, fromPosition.getNotation());
     }
